[break] Delete deprecated APIs
