+++ conjure-java-undertow-runtime/src/main/java/com/palantir/conjure/java/undertow/runtime/ConjureSerializerRegistry.java
@@ -19,7 +19,6 @@ package com.palantir.conjure.java.undertow.runtime;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ObjectArrays;
 import com.google.common.reflect.TypeToken;
-import com.palantir.conjure.java.undertow.lib.SerializerRegistry;
 import com.palantir.logsafe.Preconditions;
 import com.palantir.logsafe.SafeArg;
 import com.palantir.logsafe.exceptions.SafeIllegalArgumentException;
