--- conjure-java-undertow-runtime/src/test/java/com/palantir/conjure/java/undertow/demo/MyService.java
@@ -1,42 +1,0 @@
-/*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.conjure.java.undertow.demo;
-
-import java.time.OffsetDateTime;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-
-/** Sample API as generated by conjure-undertow. */
-public interface MyService {
-
-    /**
-     * Increments the given base date by the number of hours.
-     */
-    OffsetDateTime incrementTime(OffsetDateTime base, int numHours);
-
-    /**
-     * Returns a map indicating for each of the given dates whether they are a Sunday or not.
-     */
-    Map<OffsetDateTime, Boolean> isSunday(List<OffsetDateTime> dates);
-
-    /** Returns the trace ID made available to the server implementation. */
-    String returnTrace();
-
-    /** Returns a string iff {@code shouldReturnString == true}. */
-    Optional<String> maybeString(boolean shouldReturnString);
-}
