+++ conjure-java-undertow-runtime/src/test/java/com/palantir/conjure/java/undertow/runtime/BearerTokenLoggingTest.java
@@ -19,8 +19,7 @@ package com.palantir.conjure.java.undertow.runtime;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.palantir.conjure.java.undertow.HttpServerExchanges;
-import com.palantir.conjure.java.undertow.lib.Attachments;
-import com.palantir.conjure.java.undertow.lib.internal.Auth;
+import com.palantir.conjure.java.undertow.lib.ServiceContext;
 import io.undertow.server.HttpHandler;
 import io.undertow.server.HttpServerExchange;
 import io.undertow.server.handlers.CookieImpl;
@@ -60,6 +59,10 @@ public class BearerTokenLoggingTest {
     private static final String SESSION_ID = "3fc663d4-3e48-4ded-ba4e-d78af98b8363";
     private static final String TOKEN_ID = "a459b4a1-5089-4fe0-8655-d5dfd9b2b7fd";
 
+    private static final ServiceContext CONTEXT = ConjureContext.builder()
+            .serializerRegistry(ConjureSerializerRegistry.getDefault())
+            .build();
+
     private AtomicReference<Runnable> delegateRunnable = new AtomicReference<>();
     private HttpHandler delegate = request -> delegateRunnable.get().run();
     private HttpServerExchange exchange = HttpServerExchanges.createStub();
@@ -73,7 +76,7 @@ public class BearerTokenLoggingTest {
         exchange = HttpServerExchanges.createStub();
         exchange.setRequestMethod(Methods.GET);
         handler = new LoggingContextHandler(httpServerExchange -> {
-            Auth.header(httpServerExchange);
+            CONTEXT.authHeader(httpServerExchange);
             delegate.handleRequest(httpServerExchange);
         });
     }
@@ -102,7 +105,7 @@ public class BearerTokenLoggingTest {
     @Test
     public void testCookieAuth() throws Exception {
         handler = new LoggingContextHandler(httpServerExchange -> {
-            Auth.cookie(httpServerExchange, "PALANTIR_TOKEN");
+            CONTEXT.authCookie(httpServerExchange, "PALANTIR_TOKEN");
             assertThat(MDC.get("userId")).isEqualTo(USER_ID);
             assertThat(MDC.get("sessionId")).isEqualTo(SESSION_ID);
             assertThat(MDC.get("tokenId")).isNull();
